Create save the value of an encoder in Arduino after power off,
In this example, we are using the EEPROM library to store the encoder value in the first address (address 0) of the EEPROM. In setup(), we retrieve the encoder value from the EEPROM using the EEPROM.read() function. In loop(), we update the encoder value and print it to the serial monitor as before, but we also save it to the EEPROM using the EEPROM.write() function. We add a delay to prevent EEPROM wear, as writing to the EEPROM too often can reduce its lifespan.